perbaikan

+++ resources/views/admin/damaged/components/actions.blade.php
@@ -1,4 +1,6 @@
 <div class="btn-group">
     <button class="btn btn-sm btn-primary " onclick="showDamaged({{ $damaged->id }})">Lihat</button>
-    <button class="btn btn-sm btn-danger " onclick="deleteDamaged({{ $damaged->id }})">Hapus</button>
+    @if (Auth::user()->role == 'Gudang')
+        <button class="btn btn-sm btn-danger " onclick="deleteDamaged({{ $damaged->id }})">Hapus</button>
+    @endif
 </div>
